Build libSDLmain using libtool.
This allows it to build as a shared library on some
platforms which for Native Client acts to fix the build
for sdl-ttf and sdl-gfx (which were previously attempting
to include SDLmain statically).

This change already seems to have been made on the 1.2
branch, although its not clear if the motivation was
exactly the same.